Update getCache.js
CATALOG_TTL changed to 1 day so users get more frequent popular lists. The same issue has been reported to the stremio team and they have done so with cinemeta. One day might be a little bit much for the server and API, but it would be totally worth it. META_TTL is fine.