Fix minor oopsie in collection gathering
The previous PR (#1217) was trying to filter out the collection ID, but we want to filter out the item's ID instead. If we're seeing metadata for "Black Phone", the collection will have both it and "Black Phone 2", and we only want to retrieve the second one.

This was the intention from the start but I named my vars wrong and ended up confusing myself. Sorry!